check that the version is installed

+++ release_to_pypi.sh
@@ -10,10 +10,12 @@ rm -Rfv build/
 # python2.5 setup.py bdist_egg
 # rm -Rfv build/
 
-for ver in 2.7 3.4; do
-    python$ver setup.py bdist_egg
-    python$ver setup.py bdist_wheel
-    rm -Rfv build/
+for version in 2.7 3.4 3.5 3.6; do
+    if [ -a "`which python$version`" ]; then
+        python$ver setup.py bdist_egg
+        python$ver setup.py bdist_wheel
+        rm -Rfv build/
+    fi
 done
 
 read -p "Upload to PyPI? [y/n/x] " yesno || exit 1
